requirements.txt
streamlit==1.54.0
google-genai==1.64.0
edge-tts==7.2.7
python-dotenv==1.2.1
num2words==0.5.14
pandas==2.3.3
requests==2.32.5
